feat(evtstream): add app-grade chat package
